% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentalDesignClass.R
\name{experimentalDesign}
\alias{experimentalDesign}
\title{Create experimental design object}
\usage{
experimentalDesign(experiment.id, filepath, strand = NULL, metadata = NULL,
  id.vars = NULL)
}
\arguments{
\item{experiment.id}{an identifier for the sample}

\item{filepath}{where the files are located}

\item{strand}{The strand the data comes from +, - , or * for unknown or irrelevant}

\item{metadata}{a data.frame containing all other variables of interest}

\item{id.vars}{a character vector specifying which variables in the metadata should be interperted as experimental conditions}
}
\description{
This function creates an experimental design object that can be used to manage data
}
