% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigWigLib.R
\name{getChromInfo}
\alias{getChromInfo}
\title{Chromosome information collection}
\usage{
getChromInfo(expDes, which.chrom = NULL, regex.chrom = NULL)
}
\arguments{
\item{expDes}{experimentalDesign object}

\item{which.chrom}{character vector of which chromosomes information should be collected on. Default value of NULL collects info for all chromosomes. Cannot be specified at the same time as regex.chrom.}

\item{regex.chrom}{can supply regex expression matching which chromosomes information should be collected on. Cannot be specified at the same time as which.chrom.}
}
\description{
Collects the chromosomal lengths used to build all bigwigs experimental design object
}
