% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/experimentalDesignClass.R
\docType{methods}
\name{getFilepaths}
\alias{getFilepaths}
\alias{getFilepaths}
\title{Accessor function for experimental design object}
\description{
Return all file paths in experimentalDesign object
}
