% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/experimentalDesignClass.R
\docType{methods}
\name{getGroupIds}
\alias{getGroupIds}
\alias{getGroupIds}
\title{Accessor function for experimental design object}
\description{
Return all strands associated with experimental data
}
