% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/experimentalDesignClass.R
\docType{methods}
\name{getIds}
\alias{getIds}
\alias{getIds}
\title{Accessor function for experimental design object}
\description{
Return all experiment.ids in experimentalDesign object
}
