% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/experimentalDesignClass.R
\docType{methods}
\name{getStrand}
\alias{getStrand}
\alias{getStrand}
\title{Accessor function for experimental design object}
\description{
Return all group ids (constructed based in shared id.vars). All entries with the same group.id are replicates for the given set of experimental conditions
}
