% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/experimentalDesignClass.R
\docType{methods}
\name{getTable}
\alias{getTable}
\alias{getTable}
\title{Accessor function for experimental design object}
\description{
Return full experimental design table.
}
