% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/bigWigLib.R
\docType{methods}
\name{seperateBinnedBwElements}
\alias{seperateBinnedBwElements}
\alias{seperateBinnedBwElements}
\title{Split and clone binnedBigWig}
\usage{
\S4method{seperateBinnedBwElements}{binnedBigWig}(binnedBw)
}
\arguments{
\item{binnedBw}{binnedBigWig object}
}
\description{
Creates seperate binnedBigWig objects for each element of @counts list in masted binnedBigWig object.
}
