% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/bigWigLib.R
\docType{methods}
\name{splitBw}
\alias{splitBw}
\alias{splitBw}
\title{Split Data}
\arguments{
\item{binnedBw}{binnedBigWig object}

\item{f}{value to split binnedBigWig on}

\item{to.matrix}{LOGICAL. Should the output be converted from data.table to matrix form.}
}
\description{
Splits data from binnedBigWig into list by a column in the experimental design table.
}
