% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/experimentalDesignClass.R
\docType{methods}
\name{subsetExperiments}
\alias{subsetExperiments}
\alias{subsetExperiments}
\title{Subset experiments}
\description{
Method for producing filtered experimental design object
}
