% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigWigLib.R
\name{sumBwOverGR}
\alias{sumBwOverGR}
\title{Imports bigWig(s)}
\usage{
sumBwOverGR(bins, expDes, nthreads = 1)
}
\arguments{
\item{bins}{GRangesList object}

\item{expDes}{experimentalDesign object}

\item{nthreads}{number of cores to use to import data (not implemented)}
}
\description{
Get read counts in bw regions for given list of genomic ranges. If a strand is specified for bins, only bigwigs matching that strand will be queried. If "*" strand, both "+" and "-" strands will be queried
}
